+++ CHANGES.rst
@@ -2,8 +2,8 @@
  Changelog
 ===========
 
-1.12 (unreleased)
-=================
+1.11.1 (2017-01-07)
+===================
 
 - Fix an error in calendar display when rendering a user defined field which is
   only defined on IEvent but not on IRecurringEvent.